fixed structural problem
