+++ package.json
@@ -17,10 +17,8 @@
     "fullcalendar": "^3.0.1",
     "jquery": "^3.1.1",
     "vue": "^2.0.1",
-    "vue-full-calendar": "0.0.3",
     "vue-chartjs": "^2.2.0",
-    "vue-router": "^2.0.0",
-    "vuefire": "^1.3.0"
+    "vue-router": "^2.0.0"
   },
   "devDependencies": {
     "autoprefixer": "^6.4.0",
