+++ src/components/EventAction.vue
@@ -12,62 +12,18 @@ export default {
 		return {
 			createEventTitle: 'Create Event',
 			manageEventTitle: 'Manage Your Event',
-			events: this.$root.$firebaseRefs.events,
 			event: {}
 		}
 	},
 	methods: {
 		createEvent: function(){
-			if (this.isInputValid()) {
-				const _this = this
-				this.events.once("value").then((snapshot) => {
-					for (var key in snapshot.val()){
-						if (snapshot.val()[key] == _this.name) {
-							swal(
-								'Event create failed',
-								'The event name has already been taken, please try again',
-								'error'
-							)
-							return
-						}
-						continue
-					}
-					let eventKey = _this.events.push(_this.name).key
-					_this.event[eventKey] = {
-						name: _this.name,
-						size: {
-							max: 10,
-							min: 1
-						}
-					}
-					_this.$root.$firebaseRefs.root.update(_this.event)
-					swal(
-						'Event created',
-						'Hit OK to proceed to your event page',
-						'success'
-					).then(function() {
-						_this.$router.push({ name: 'event', params : { id: eventKey }})
-					})
-				})
+			if (this.isInputValid()){
+				this.$store.dispatch('createEvent', this.name)
 			}
 		},
 		loadEvent: function(){
 			if (this.isInputValid()) {
-				const _this = this
-				this.events.once("value").then((snapshot) => {
-					for (var key in snapshot.val()){
-						if (snapshot.val()[key] == _this.name) {
-							_this.$router.push({ name: 'event', params : { id: key }})
-							return
-						}
-						continue
-					}
-					swal(
-						'Event not found',
-						'The event name entered does not exist, please try again',
-						'error'
-					)
-				})
+				this.$store.dispatch('loadEvent', this.name)
 			}
 		},
 		isInputValid: function(){
