+++ src/main.js
@@ -1,3 +1,14 @@
-import root from './root';
+import Vue from 'vue'
+import VueRouter from 'vue-router'
+import routes from './routes'
 
-root.$mount('#app');
+Vue.use(VueRouter);
+
+const router = new VueRouter({
+  routes,
+  mode: 'history'
+});
+
+const app = new Vue({
+  router,
+});
